support object and arrays

+++ src/lib/writer/index.js
@@ -6,7 +6,7 @@ import path from 'path';
 
 const target = 'writer';
 
-export default function({log, item, content}){
+export default function({log, item, content, simulate}){
   const def = Q.defer();
   
   let dir = path.dirname(item.fullname);
@@ -17,14 +17,18 @@ export default function({log, item, content}){
   let rel = path.join(item.relative, item.name);
 
   log.verbose(target, `writing ${rel}`);
-  fs.writeFile(item.fullname, content, (err) => {
-    if (err){
-      log.error(target, err);
-      return def.reject(err);
-    }
+  if (!simulate){
+    fs.writeFile(item.fullname, content, (err) => {
+      if (err){
+        log.error(target, err);
+        return def.reject(err);
+      }
 
-    return def.resolve(item);
-  });
+      return def.resolve(item);
+    });
+  } else {
+    def.resolve(item);
+  }
 
   return def.promise;
 }
